scripts/user-luxmark3: no division by zero pdf clamp
It prevents a divizion by zero to happen when pdf clamp value is set to zero.

Big thanks to Matt Arsenault for the precious help provided to track down this bug.

Since LLVM commit 29a2b20ab363bcc0b9573e358a5ad12c0eddca86 from LLVM 11,
this bug breaks the rendering when both -cl-finite-math-only
and -cl-unsafe-math-optimizations are set.

This happens when -cl-fast-relaxed-math is set.

The result of the bad division by zero is used in a multiplication that
is optimized in a way the multiplication result itself becomes wrong.

The patch fixes the issue when LuxMark 3.1 is running on Mesa Clover
or on AMD ROCm OpenCL platform, as both platforms use LLVM amdgcn
bytecode generation.

The LLVM FMUL optimization that uncovered the bug is there:

https://github.com/llvm/llvm-project/blob/29a2b20ab363bcc0b9573e358a5ad12c0eddca86/llvm/lib/CodeGen/SelectionDAG/DAGCombiner.cpp#L12508-L12509

The broken computation dividing something by pdf clamp even if it's zero is there:

https://github.com/LuxCoreRender/LuxCore/blob/luxmark_v3.1/include/slg/engines/pathocl/kernels/pathocl_kernels_micro.cl#L642-L643

The pdf clamp value is set to zero there:

https://github.com/LuxCoreRender/LuxCore/blob/luxmark_v3.1/src/slg/engines/pathocl/pathocl.cpp#L121

The patches rewrites this:

  min(1.f, something / PARAM_PDF_CLAMP_VALUE)

into this:

  min(1.f, something / (PARAM_PDF_CLAMP_VALUE == 0.f ? something : PARAM_PDF_CLAMP_VALUE))

Since something / something == 1.f, doing this never divides by zero, while doing that
would still divide by zero (GPUs may still compute the unused code even if not using it):

  min(1.f, PARAM_PDF_CLAMP_VALUE == 0.f ? 1.f : something / PARAM_PDF_CLAMP_VALUE))

This patch fixes:

- I Love Compute: https://gitlab.com/illwieckz/i-love-compute/-/issues/16
- I Love Compute: https://gitlab.com/illwieckz/i-love-compute/-/issues/25
- LuxCoreRender: https://forums.luxcorerender.org/viewtopic.php?t=4600
- LLVM: https://github.com/llvm/llvm-project/issues/54947
- Clover: https://gitlab.freedesktop.org/mesa/mesa/-/issues/5671
- ROCm: https://github.com/RadeonOpenCompute/ROCm/issues/1828